Payroll Detail Entry now ok

+++ IspDataDb/dbo/Tables/PayrollDetail.sql
@@ -1,10 +1,12 @@
 ﻿CREATE TABLE [dbo].[PayrollDetail] (
     [IdNo]          INT        IDENTITY (1, 1) NOT NULL,
-    [PayrollIdNo]   TINYINT    NULL,
+    [PayrollIdNo]   SMALLINT   NULL,
     [EmployeeIdNo]  INT        NULL,
     [DateTimeStamp] ROWVERSION NULL,
     CONSTRAINT [PK_PayrollDetail] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
